Added confirmation for edit/add/delete

+++ src/songLib/view/SongLibController.java
@@ -2,9 +2,13 @@ package songLib.view;
 
 import javafx.event.ActionEvent;
 import java.util.List;
+import java.util.Optional;
+
 import javafx.fxml.FXML;
 import javafx.scene.control.Alert;
 import javafx.scene.control.Button;
+import javafx.scene.control.ButtonType;
+import javafx.scene.control.DialogPane;
 import javafx.scene.control.ListView;
 import javafx.scene.control.TextField;
 import javafx.scene.control.Alert.AlertType;
@@ -12,6 +16,7 @@ import javafx.stage.Stage;
 import javafx.scene.layout.GridPane;
 import javafx.scene.layout.HBox;
 import javafx.scene.layout.VBox;
+import javafx.scene.text.Text;
 import songLib.app.*;
 
 public class SongLibController {
@@ -43,8 +48,6 @@ public class SongLibController {
 					displayInfo(songList.getSelectionModel().getSelectedItem());
 					if (editMode) toggleEdit(null);
 				});
-		
-		
 	}
 	
 	public void start(Stage primaryStage) {
@@ -76,6 +79,7 @@ public class SongLibController {
 
 	public void addCommand(ActionEvent e) {
 		String[] details = {createName.getText().trim(), createArtist.getText().trim(), createAlbum.getText().trim(), createYear.getText().trim()};
+		if (!confirm("Are you sure you want to add song: " + details[0] + " " + details[1])) return;
 		if (!verifyDetails(details))
 			return;
 		Song newAddition = library.addSong(details);
@@ -89,6 +93,7 @@ public class SongLibController {
 	}
 	
 	public void deleteCommand(ActionEvent e) {
+		if (!confirm("Are you sure you want to delete song: " + songList.getSelectionModel().getSelectedItem())) return;
 		int currIndex = songList.getSelectionModel().getSelectedIndex();
 		library.deleteSong(songList.getSelectionModel().getSelectedItem());
 		songList.getSelectionModel().select(currIndex);
@@ -143,12 +148,31 @@ public class SongLibController {
 	}
 	
 
+	public boolean confirm(String content) {
+		Alert alert = new Alert(AlertType.CONFIRMATION);
+		alert.initOwner(primaryStage);
+		alert.setTitle("Confirm");
+		alert.setHeaderText(content);
+		Text t = new Text("This is a text sample");
+		t.setStyle("-fx-font-family: Arial");
+		DialogPane dialogPane = alert.getDialogPane();
+		dialogPane.setContent(t);
+		Optional<ButtonType> result = alert.showAndWait();
+		if(!result.isPresent() || result.get() != ButtonType.OK) {
+			return true;
+		} 
+		return false;
+		
+		
+	}
+	
 	
 	public void submitEdits(ActionEvent e) {
 		String[] details = {songName.getText().trim(), songArtist.getText().trim(), songAlbum.getText().trim(), songYear.getText().trim()};
 		if (!verifyDetails(details))
 			return;
 		Song edited = songList.getSelectionModel().getSelectedItem();
+		if (!confirm("Change song details from: " + edited + " to " + details[0] + details[1])) return; 
 		edited.setName(songName.getText());
 		edited.setArtist(songArtist.getText());
 		edited.setAlbum(songAlbum.getText());
@@ -195,11 +219,6 @@ public class SongLibController {
 		alert.showAndWait();
 
 		return false;
-
-		
-		
-		
-		
 	}
